Convert generated inline type name to model name before generating unique name.
This avoids a possible clash between an inline type and another inline or defined
type with a name that only differs in naming style. I.e. inline object
parameter "person" and defined type "Person" will result in the same class name
when camelized, leading to wrong type definitions for "person".